+++ examples/bar/defaultbar.py
@@ -1,0 +1,5 @@
+from quo.bar import Bar
+
+bar = Bar("I`m a Bar")
+
+bar.draw()