+++ examples/bar/left-alignedbar.py
@@ -1,0 +1,5 @@
+from quo.bar import Bar
+
+bar = Bar("I am a left aligned bar")
+
+bar(align="left")