+++ examples/bar/right-alignedbar.py
@@ -1,0 +1,5 @@
+from quo.bar import Bar
+
+bar = Bar("I am a right aligned bar")
+
+bar(align="right")