+++ examples/bar2.py
@@ -1,5 +1,0 @@
-from quo.console import Console
-
-console = Console()
-
-console.bar("I am a styled bar", style="fg:green bg:brown")