+++ examples/bar3.py
@@ -1,5 +1,0 @@
-from quo.console import Console
-
-console = Console()
-
-console.bar("I am a left aligned bar", align="left")