+++ examples/dialogs/messagebox.py
@@ -2,10 +2,10 @@
 """
 Example of a message box window.
 """
-from quo.dialog import _MessageBox as MessageBox
+from quo.dialog import MessageBox
 
 MessageBox(
-        title="Message pop up window",
+        title="Message window",
         text="Do you want to continue?\nPress ENTER to quit."
-        ).run()
+        )