+++ examples/dialogs/passwordbox.py
@@ -2,7 +2,6 @@
 """
 Example of an password input dialog.
 """
-from quo import print
 from quo.dialog import InputBox
 
 result = InputBox(