+++ examples/dialogs/radiolist2.py
@@ -1,0 +1,10 @@
+from quo.dialog import RadiolistBox
+
+RadiolistBox( 
+    title="RadioList dialog example", 
+    text="Which breakfast would you like ?", 
+    values=[ 
+        ("breakfast1", "Eggs and beacon"), 
+        ("breakfast2", "French breakfast"), 
+        ("breakfast3", "Equestrian breakfast") 
+    ])