+++ examples/key_binding/kb.py
@@ -16,6 +16,3 @@ def _(event):
 #session.prompt(">> ")
 
 
-
-if __name__ == "__main__":
-    _(event)