+++ examples/prompts/int-validator.py
@@ -1,7 +1,7 @@
-#!/usr/bin/env python
 
 """Simple example of an input validator"""
 
 from quo import prompt
 
-prompt("Type a integer: ", type=int)
+prompt("Type a integer: ", type=float)
+p