+++ examples/prompts/placeholder.py
@@ -6,7 +6,7 @@ from quo import echo
 from quo.prompt import Prompt
 from quo.text import Text
 
-session = Prompt(placeholder=Text('style fg="gray">(please type something)</style>'))
+session = Prompt(placeholder=Text('<style fg="gray">(please type something)</style>'))
 
 if __name__ == "__main__":
     answer = session.prompt(">> ")