+++ examples/rule.py
@@ -1,5 +1,0 @@
-from quo.console import Console
-
-console = Console()
-
-console.rule(style="red")