+++ examples/rule/default-rule.py
@@ -1,0 +1,4 @@
+from quo.rule import Rule
+
+rule = Rule()
+rule.draw()