+++ examples/rule2.py
@@ -1,5 +1,0 @@
-from quo.console import Console
-
-console = Console()
-
-console.rule(animated=True)